update to add array lengths at starts of arrays
This commit will be updated in-place and is not complete.

The Ouroboros handshake CBOR decoder seems to expect the array length to
be put somewhere, but I've not quite worked out how to get to be accepted
all the way through quite yet. At some point, the placement and
formatting of the array lengths in the CBOR will be worked out.