ignore mini-protocols apart from trace forwarding
The protocol number is kept in a field of the SDU. This defines an
enumeration for the 3 used during the main runtime, while also giving
access to the field as a uint16_t within a union. It needs to be done,
though there are momentarily other things going awry preventing its
immediate effects from being seen.